// src/core/types/formikValues/BankProtal/BankProtal.values.ts
//BankProtalResult
export interface IBankProtalResult {
  MID: string | null;
  TerminalId: string | null;
  RefNum: string | null;
  ResNum: string | null;
  Amount: string | null;
  Wage: string | null;
  Rrn: string | null;
  SecurePan: string | null;
  HashedCardNumber: string | null;
  Status: string | null;
  State: string | null;
  TraceNo: string | null;
}